+++ app/src/androidTest/java/com/mindgarden/mindgarden/DiaryDatabaseTest.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import android.util.Log
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
-import com.mindgarden.mindgarden.data.model.entity.Diary
-import com.mindgarden.mindgarden.data.repository.database.MindGardenDatabase
+import com.mindgarden.mindgarden.data.db.MindGardenDatabase
+import com.mindgarden.mindgarden.data.db.entity.Diary
 import com.mindgarden.mindgarden.data.repository.diaryRepo.local.DiaryDao
 import org.junit.After
 import org.junit.Before
@@ -28,34 +28,5 @@ class DiaryDatabaseTest {
     @Throws(IOException::class)
     fun closeDb() {
         db.close()
-    }
-
-    @Test
-    fun diaryTest() {
-        val date = Calendar.getInstance().time
-        val diary = Diary(
-            0,
-            date,
-            "diary test",
-            1,
-            "good day", null
-        )
-        diaryDao.insert(diary).subscribe(
-            {
-                Log.d("DiaryDatabaseTest", "Success insert")
-            },
-            {
-                Log.e("DiaryDatabaseTest", "Error : $it")
-            }
-        )
-
-        diaryDao.loadDiaries(date).subscribe(
-            {
-                Log.d("DiaryDatabaseTest", "Success load diary : ${it.size}")
-            },
-            {
-                Log.e("DiaryDatabaseTest", "Error : $it")
-            }
-        )
     }
 }
