+++ app/src/main/java/com/mindgarden/mindgarden/data/db/Converters.kt
@@ -1,27 +1,12 @@
-package com.mindgarden.mindgarden.data.repository.common
+package com.mindgarden.mindgarden.data.db
 
 import androidx.room.TypeConverter
 import java.time.LocalDateTime
-import java.util.*
-
 
 object Converters {
-
     @TypeConverter
     @JvmStatic
-    fun fromTimestampToDate(timestamp: Long?): Date? {
-        return timestamp?.let { Date(it) }
-    }
-
-    @TypeConverter
-    @JvmStatic
-    fun dateToTimestamp(date: Date?): Long? {
-        return date?.time
-    }
-
-    @TypeConverter
-    @JvmStatic
-    fun toDate(dateString: String?): LocalDateTime? {
+    fun toLocalDateTime(dateString: String?): LocalDateTime? {
         return if (dateString == null) {
             null
         } else {
