+++ app/src/main/java/com/mindgarden/mindgarden/data/db/MindgardenDatabase.kt
@@ -1,23 +1,19 @@
-package com.mindgarden.mindgarden.data.repository.database
+package com.mindgarden.mindgarden.data.db
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.mindgarden.mindgarden.data.model.entity.Diary
-import com.mindgarden.mindgarden.data.model.entity.Mind
-import com.mindgarden.mindgarden.data.model.entity.User
-import com.mindgarden.mindgarden.data.repository.common.Converters
+import com.mindgarden.mindgarden.data.db.entity.Diary
+import com.mindgarden.mindgarden.data.db.entity.Mind
 import com.mindgarden.mindgarden.data.repository.diaryRepo.local.DiaryDao
 import com.mindgarden.mindgarden.data.repository.gardenRepo.local.dao.GardenDao
-import com.mindgarden.mindgarden.data.repository.userRepo.local.dao.UserDao
 
-@Database(entities = [Mind::class, Diary::class, User::class], version = 1, exportSchema = false)
+@Database(entities = [Mind::class, Diary::class], version = 1, exportSchema = false)
 @TypeConverters(Converters::class)
 abstract class MindGardenDatabase : RoomDatabase() {
     abstract fun gardenDao(): GardenDao
-    abstract fun userDao(): UserDao
     abstract fun diaryDao(): DiaryDao
 
     companion object {
