+++ app/src/main/java/com/mindgarden/mindgarden/data/db/entity/Mind.kt
@@ -1,5 +1,6 @@
-package com.mindgarden.mindgarden.data.model.entity
+package com.mindgarden.mindgarden.data.db.entity
 
+import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import java.time.LocalDateTime
@@ -7,9 +8,9 @@ import java.time.LocalDateTime
 @Entity
 data class Mind(
     @PrimaryKey(autoGenerate = true)
-    val idx: Long?,
+    @ColumnInfo(name = "idx")
+    val idx: Long,
     val treeIdx: Int,
-    val gardenDate: LocalDateTime, // e.g. 2021-09-01T00:00
-    val mindDate: LocalDateTime, // e.g. 2021-09-22T07:46:14.231
     val location: Int,
+    val date: LocalDateTime, // 2021-09-22T07:46:14.231
 )
