+++ app/src/main/java/com/mindgarden/mindgarden/data/repository/diaryRepo/DiaryRepository.kt
@@ -1,6 +1,6 @@
 package com.mindgarden.mindgarden.data.repository.diaryRepo
 
-import com.mindgarden.mindgarden.data.model.entity.Diary
+import com.mindgarden.mindgarden.data.db.entity.Diary
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Flowable
 import io.reactivex.rxjava3.core.Single
