+++ app/src/main/java/com/mindgarden/mindgarden/data/repository/gardenRepo/GardenRepository.kt
@@ -1,6 +1,6 @@
 package com.mindgarden.mindgarden.data.repository.gardenRepo
 
-import com.mindgarden.mindgarden.data.model.entity.Mind
+import com.mindgarden.mindgarden.data.db.entity.Mind
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
 import java.time.LocalDateTime
