+++ app/src/main/java/com/mindgarden/mindgarden/data/repository/gardenRepo/local/dao/GardenDao.kt
@@ -2,7 +2,7 @@ package com.mindgarden.mindgarden.data.repository.gardenRepo.local.dao
 
 import androidx.room.Dao
 import androidx.room.Query
-import com.mindgarden.mindgarden.data.model.entity.Mind
+import com.mindgarden.mindgarden.data.db.entity.Mind
 import com.mindgarden.mindgarden.data.repository.common.BaseDao
 import io.reactivex.rxjava3.core.Single
 import java.time.LocalDateTime
@@ -10,6 +10,6 @@ import java.time.LocalDateTime
 @Dao
 abstract class GardenDao : BaseDao<Mind> {
 
-    @Query("SELECT * from mind WHERE gardenDate=:date")
+    @Query("SELECT * from mind WHERE date=:date")
     abstract fun getGarden(date: LocalDateTime): Single<List<Mind>>
 }
