+++ app/src/main/java/com/mindgarden/mindgarden/ui/inventory/model/InventoryMind.kt
@@ -1,11 +1,10 @@
 package com.mindgarden.mindgarden.ui.inventory.model
 
-import com.mindgarden.mindgarden.data.model.entity.Mind
+import com.mindgarden.mindgarden.data.db.entity.Mind
 import java.time.LocalDateTime
 
 data class InventoryMind(
     val idx: Long? = null,
-    val gardenDate: LocalDateTime,
     val date: LocalDateTime,
     val location: Int,
     var treeIdx: Int? = null, // convert DrawableRes
@@ -14,16 +13,14 @@ data class InventoryMind(
     companion object {
         fun Mind.convertInventoryMind() = InventoryMind(
             idx = this.idx,
-            gardenDate = this.gardenDate,
-            date = this.mindDate,
+            date = this.date,
             location = this.location,
             treeIdx = this.location,
             type = GardenType.PLANTED)
 
         fun InventoryMind.convertMind() = Mind(
-            idx = null,
-            gardenDate = this.gardenDate,
-            mindDate = this.date,
+            idx = this.idx ?: 0,
+            date = this.date,
             location = this.location,
             treeIdx = this.treeIdx ?: -1
         )
