+++ app/src/main/java/com/mindgarden/mindgarden/util/ext/LocalDateTimeExt.kt
@@ -4,5 +4,5 @@ import java.time.LocalDateTime
 
 fun now(): LocalDateTime = LocalDateTime.now()
 
-fun LocalDateTime.getGardenDate(): LocalDateTime =  LocalDateTime.of(
+fun LocalDateTime.toGardenDate(): LocalDateTime =  LocalDateTime.of(
     this.year, this.month, 1, 0, 0,0,0)
